Fixed armoring a bit
Also did so that the connection is a "dummy", meaning it is not
connected to a server but can behave like it does. Next up will be
representing the correct values on different bodyparts and build methods
needed to store and check these values. And also fix a disconnect gui.