Added some more functionality.
Fixed armorpoint handling and added a waiting gui, the armorpoint
handling in Module is just a dummy to check for bugs atm. But maybe make
it that the server grants an maximum points of armor to Module-client
that the user fills as he/she want to and then send back the results?
(Like Head should have 3 points and so on...) I am thinking about
minimizing networktraffic.